extract shared constants, avoid circular imports
