+++ src/FreeScribe.client/UI/MainWindowUI.py
@@ -1,7 +1,7 @@
 import tkinter as tk
 from tkinter import ttk
 import UI.MainWindow as mw
-from UI.SettingsWindow import FeatureToggle
+from UI.SettingsConstant import FeatureToggle
 from UI.SettingsWindowUI import SettingsWindowUI
 from UI.MarkdownWindow import MarkdownWindow
 from utils.file_utils import get_file_path
